Retain old ctor as deprecated and provide new ctor with target
