+++ src/main/java/com/jenkinsci/plugins/badge/action/BadgeSummaryAction.java
@@ -41,9 +41,17 @@ public class BadgeSummaryAction extends AbstractBadgeAction {
 
     private static final Logger LOGGER = Logger.getLogger(BadgeSummaryAction.class.getName());
 
+    /**
+     * @deprecated Use {@link BadgeSummaryAction#BadgeSummaryAction(String, String, String, String, String, String, String)} instead.
+     */
+    @Deprecated
+    public BadgeSummaryAction(String id, String icon, String text, String cssClass, String style, String link) {
+        super(id, icon, text, cssClass, style, link);
+    }
+
     public BadgeSummaryAction(
             String id, String icon, String text, String cssClass, String style, String link, String target) {
-        super(id, icon, text, cssClass, style, link);
+        super(id, icon, text, cssClass, style, link, target);
     }
 
     @Whitelisted
